update %p to use unsigned long instead of uintptr_t

+++ ft_printf/ft_printf.h
@@ -23,8 +23,8 @@ int	ft_putchar(int c);
 int	ft_putstr(const char *str);
 int	ft_putuint(unsigned int n);
 int	ft_putint(int n);
-int	ft_puthex(uintptr_t n, char c);
-int	ft_putptr(uintptr_t ptr);
+int	ft_puthex(unsigned long n, char c);
+int	ft_putptr(unsigned long ptr);
 int	ft_ulen(unsigned int n);
 int	ft_ilen(int n);
 #endif
